ima: move IMA components into their own module
