+++ keylime/agentstates.py
@@ -6,9 +6,9 @@ Copyright 2021 IBM Corporation
 
 import threading
 
-from keylime.ima_ast import get_FF_HASH, get_START_HASH
+from keylime.ima.ima_ast import get_FF_HASH, get_START_HASH
 from keylime.common.algorithms import Hash
-from keylime.ima_file_signatures import ImaKeyrings
+from keylime.ima.ima_file_signatures import ImaKeyrings
 
 class TPMState():
     """ TPMState models the state of the TPM's PCRs """
